Add webapp sample
We're adding a simple Flask application meant to showcase MongoDB.
It defines a "store" resource, allowing CRUD operations as well
as text and geospatial searches.

Check the readme for some examples and an installation guide.

In order to ensure that the application continues to work properly,
we've added some functional tests. Use tox to run these tests along
with flake8 linter tests.